requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
redis
psycopg2-binary
sec-api
supabase
python-multipart
polars
pandas
openpyxl
beautifulsoup4
pypdf
pdfplumber
lxml
pyarrow
fastexcel
pytesseract
Pillow
paddleocr
paddlepaddle
pdf2image
opencv-python-headless
playwright
numpy
h5py
pypdfium2
google-generativeai
google-genai
cryptography
requests
python-jose
sentence-transformers
pyspark
delta-spark
mlflow
